fastapi/requirements.txt
mysql-connector-python
fastapi
uvicorn
gunicorn
